warp-tls: ignore socket errors while sending close_notify
(fixes #640)

RFC 5246 section 7.2.1 says this about closure alerts:

> The client and the server must share knowledge that the connection is ending
> in order to avoid a truncation attack.  Either party may initiate the
> exchange of closing messages.
>
> [...]
>
> Unless some other fatal alert has been transmitted, each party is required to
> send a close_notify alert before closing the write side of the connection.
> The other party MUST respond with a close_notify alert of its own and close
> down the connection immediately, discarding any pending writes.  It is not
> required for the initiator of the close to wait for the responding
> close_notify alert before closing the read side of the connection.

This `close_notify` message is sent by `TLS.bye`, and it used to generate an
exception when the receiving side closes the connection before waiting for
`close_notify`. But this scenario is perfectly legal according to the RFC, so
we should ignore `ConnectionClosedByPeer` exception that happens during
`TLS.bye`.